Statically link model builders in CLI (#727)
To deal with the [current limitation of cargo dist](https://github.com/axodotdev/cargo-dist/issues/179), which doesn't support distributing dynamic libs, we statically link model-builders.

We still retain the capability to build and link model-builders dynamically. This is needed for tests in core packages, which cannot have static dependency on *-model-builders (will create a circular dependency).