WIP (we get the value, but not using it yet)
Implement context-based default values for relationships

This allows to specify default values such as:
```exo
type Todo {
  ...
  user: User = AuthContext.id
}
```

This is useful, for example, to automatically set the user of a todo
to the current user.

This, in turn, will help specify input validation rules to reject
a usee specifying another user when updating a todo.

A major change required to implement this is to make several traits
and functions async, since extracting a value from context is an
async operation.