Default value from context (#731)
Implement context-based default values for relationships. This PR implements only the parsing part and some way through resolver, but doesn't complete it. We need to unify arguments types (currently, we have JSON `Value`, GraphQL `ConstValue`, which makes code unnecessarily complex). But we are merging this PR, since this already has significant changes around async nature and it will be harder to merge when we implement the unified argument type.

This allows to specify default values such as:
```exo
type Todo {
  ...
  user: User = AuthContext.id
}
```

This is useful, for example, to automatically set the user of a todo
to the current user.

This, in turn, will help specify input validation rules to reject
a user specifying another user when updating a todo.

A major change required to implement this is to make several traits
and functions async, since extracting a value from context is an
async operation.