Unified val type (#732)
So far we used a combination of `serde_json::Value` and `async_graphql_value::ConstValue` to represent
arguments, context objects, etc. This caused some friction when, for example, assigning a default
value from context if an argument is not provided (two types don't match, forcing a conversion, but
in that process making one object a references and the other a value, which made code complex).

In this PR, we introduce our own `Val` type. Currently, it is similar to `async_graphql_value::ConstValue`,
but we will extend it in the future to support more types (e.g. `Int` and `Float` (instead of just
`Number`), `Instant` and `LocalTime` instead of just `String`, etc).

In the process:
- cleaned up code for RequestContext and related types.
- Remove dependency on `async-graphql-value` from postgres/deno resolvers