Support a better project directory structure (#739)
Earlier, we created exo_ir files in the same directory as the exo file, which causes some confusion on what needs to be committed. Also, we didn't have a good place to put integration tests. To solve these issues, we now expect the following structure:

```
src/
  index.exo
  code.ts
  ...
tests/
  *.init
  *.gql
  *.exotest
  ...
target/ (created by exo build)
  index.exo_ir
```

This PR implements this structure and updates the structure of integration-tests to match it.