Produce context.d.ts file in <project-root>/generated (#742)
This file is shared by all modules, so it needs to be in a common location. By putting in "generated" (which we add to .gitignore), we are free to regenerate as needed to ensure that those types are in sync with context types in the exo files.

This also fixes the infinite watch look due to regenerated context.d.ts files.