Improve the 'exo dev' experience for migration
Earlier, when there is a model change that requires a mighration, we just alerted the user and let them fix it outside of the 'exo dev' flow.

In this change, we offer a migration option. We warn if there are destructive changes and upon confirmation, carry the migrations.

Also, use the 'inquire' crate for better UI (and partity with 'yolo').