Improve the 'exo dev' experience for migration (#752)
Earlier, when a model change required a migration, we just alerted the user and let them fix it outside of the 'exo dev' flow.

In this change, we offer a migration option. We warn if there are destructive changes and, upon confirmation, carry the migrations.

Also, use the 'inquire' crate for better UI (and parity with 'yolo').