Async cli (#755)
Earlier, several places created a tokio runtime and used `block_on`. This created issues such as creating runtime within runtime and executing functions without having a runtime. Specifically, `exo yolo` was broken due to runtime-in-runtime.

This PR makes the whole cli a `#[tokio::main]` process obviating the need to create more runtimes and free-ing ourselves from ensuring the execution of a task in a runtime context.

We still have a `block_on` in `process_script` that creates a Deno bundle for the reasons mentioned in the comment.