Push the knowledge of database schema in exo_sql (#765)
So far, we dereference `PhysicalTable` and `PhysicalColumn` in the resolver and pass those references in `AbstractSelect` etc. This makes exo-sql work with limited knowledge (for example, doesn't quite know about referenced columns). This, in turn, puts the burden on the resolver to encode relevant information (for example, using `ColumnPath`).

In this PR, we construct a `Database` object, which contains `PhysicalTable`s. To most exo-sql APIs, we pass a reference to this `Database` and just the `TableId`s and `ColumnId`s in place of `PhyiscalTable`s and `PhysicalColumn`s. We still have `ColumnPath` (except it contains `ColumnId`s). We will deal with simplifying that separately.

Also:
* Rename Database to DatabaseClient
* Rename ColumnIdPath/Link to PhyiscalColumnPath/Link
* Introduce DatabaseSpec, TableSpec, and ColumnSpec to represent database information in a raw form.
* Removed lifetime from AbstractSelect, etc.

Fixes #763