Use CARGO_BIN_EXE_exo to locate binary for cli smoke tests
This is a more reliable way to find the exectable since the target
directory may or may not contain the build target name.

The CI adds `--target` to the build and test commands which results in
the target name (e.g. x86_64-unknown-linux-gnu) being added to the output
path. Using this environment variable means we always know we are going
to get the correct path to the executable.