Build server-aws-lambda only on Linux platforms
Since, we don't need to build the server-aws-lambda crate on Mac/Windows platforms, to save build
times (for Mac platforms) and to avoid build/test issues (for Windows platforms), we build the
server-aws-lambda crate only on Linux platforms.

Ideally, we could have included the server-aws-lambda crate in the workspace only on Linux
platforms, but until https://github.com/rust-lang/cargo/issues/5220 is fixed, we can't do that.