Add basic smoke tests for exo cli (#766)
- Run exo new and check for expected files
- Create schema and check output
- Run exo build and check ir file is created

The tests are only on unix for now since rexpect doesn't compile on Windows.

The CARGO_BIN_EXE_exo variable is used to locate the binary, which is more
reliable than working out the path ourselves - it may or may not include the target
name (e.g. x86_64-unknown-linux-gnu),  for example.