Make SubsystemLoader async and add DB startup checks
This reinstates the DB connection check which was removed when
switching to an async connection pool (see #328).

In order to call get_client() in DatabaseClient to check the
connection, the initialization functions have to be async. The
client is setup when loading the Postgres subsystem, hence the
loader also has to be async.

Some of the tests (such as the introspection tests) initialize the
Postgres subsystem without connecting to a real database, so it's
necessary to override the default in these cases, making sure the
check isn't performed.