Refactor exo-sql::Database to carry relation information
Before this change, the relationship information was (partially) kept in the GraphQL model (in `PostgresRelation`). That made writing tests in exo-sql difficult. For example, we had to create `ColumnPath`s manually in tests, and there existed the possibility of creating invalid paths.

This PR makes exo-sql self-sufficient and then uses that to clean up other parts of the code:
- Keep `ManyToOne` relations in `Database` (the other kind: `OneToMany` can be derived from it)
- Refactor `ColumnPath` to maintain invariants to avoid creating invalid paths
- Remove the `PhysicalColumnType::ColumnReference` variant since the `ManyToOne` relationship makes it redundant
- Use `RelationId` instead of `ColumnPathLink` for `Subselect` (avoids the illegal case of passing a leaf link)

Fixes #780