Replace openssl with rustls
We had two dependencies on openssl: Postgres and http-req.

For Postgres, we now use a combination of rustls, tokio-postgres-rustls,
rustls-native-certs, and rustls-pemfile crates to provide the same
functionality.

This commit drops the support for "enabling ssl but ignoring
certificate validation" (through the EXO_SSL_VERIFY env variable). This
was possible with OpenSSL, but rustls does not support it (and so far we
haven't come across a use case for it).

  Tested with:
  - local Postgres without SSL
  - Neon (with sslmode=true)
  - Yugabyte (with sslrootcert=./root.crt") to test custom root certs

For http-req, we use the rustls feature to replace openssl with rustls.
This usage was only during "exo test" (and verified that it works fine).

Fixes #798