Improve user experience when deploying to fly.io (#820)
We created fly.toml and Dockerfile in target/fly, implying that
the users shouldn't modify them. However, once deployed, users
will want to update it (for adding environment variables, updating the
version, setting the region, etc.). Identical issues apply to the
Dockerfile generated (here, modifications would be setting the timezone,
installing additional packages etc.). In other words, the generated
files are just a starting point.

In this commit, we generated the files in the project root and update
the instructions accordingly.