Make naming of env vars constants consistent (#815)
Some of the declarations of env vars had the "EXO_" prefix, some
had the "_PARAM" suffix. This commit makes them all consistent. In the
process, it also introduces constants for all `EXO_*` envs.

There is some duplication (between cli and other crates), which is
not easily avoidable since cli doesn't depend on, for example,
the server_common crate.

We leave the "OTEL_" prefix for OpenTelemetry env vars as locally defined
string literals since all of them are used in a single place.