Do not exit yolo/dev mode on ts/js errors (#832)
During building the dependency graph for Deno, we may encounter ts/js errors. We used the `Generic` error, which is currently treated as an unrecoverable error, which led to exiting the yolo/dev watch loop.

In this commit, we introduce an `ExternalResourceParsing` variant, which emits the given error message but does not exit.

Also, pass a diagnostic error message to the `ExternalResourceParsing` variant so we can show the user a helpful message.