Improve live reloading experience of playground
If `EXO_INTROSPECTION_LIVE_UPDATE` is enabled (automatically set to true in yolo and dev mode), refetch the schema every 2 seconds.

Also, take care of error situations such as a query-less model (for example, due to an empty exo file).

In the future, we can use websocket to have the server notify the client of a schema change (instead of polling), but this sets the foundation for that.

Fixes #299