Add an optional 'retries' field to exotest files (#847)
If set the test will be retried this many times if it fails until
it passes (or panics). If it still doesn't pass, the last failure
will be reported for the test.

There is a pause before each retry, starting at one second duration
and doubling each time thereafter.

Fixes #843.