Avoid duplicated `AstExpr` clones for access control
Earlier, we cloned the `AstExpr` for each access control rule based on
the availability of the expression (for example, if the `mutation` expression
os provided and not `create`/`update`/`delete`, we cloned the `mutation` expression
for each of those).

Now that we are using an arena, we avoid this by deferring the cloning
until the final bulding stage, and even then we clone only the arena index.