Use an arena to keep access control expression (#864)
This is in preparation for an upcoming PR, where we will need to make access control expressions (which are not `Clone`) part of other cloneable structs. So we introduce an arena to keep the indices (which are cloneable) in those structs.

Even though support for that functionality is the primary goal, this PR improves code by avoiding duplicated clones of `AstExpr`. Earlier, we cloned the `AstExpr` for each access control rule based on the availability of the expressions (for example, if the `mutation` expression is provided and not `create`/`update`/`delete`, we cloned the `mutation` expression for each of those). Now that we are using an arena, we avoid this by deferring the cloning until the final building stage, and even then, we clone only the arena index.