Implement JWT token generation in the playground (#863)
We now offer generating JWT tokens in the playground to simplify testing access control rules. We implement a new plugin for the playground, which asks for the JWT secret and payload, and then generate a token. The token is then added to the HTTP headers of the requests sent to the server.

Also, make the following seemingly unrelated changed need to make this functionality work:
- Change where we start the server from "0.0.0.0" to "127.0.0.1". We wanted to do this anyway, but the JWT token generation in the playground forced us to do it now (using `CryptoKey` in the browser requires to be on a "https" connection, and browsers consider "localhost" as well as "127.0.0.1 to be "secure", but not "0.0.0.0").
- Upgrade GraphiQL and related packages. Since we have started passing the `headers` props, we experienced the exact same issue as described in
https://github.com/graphql/graphiql/commit/2348641c07748691c478ac5f67032b7e9081f9cb, so we upgrade
the version to fix it.