Improve implementation of access check for nested creation (#869)
Introduce `TransactionStep::Filter` to take filter the rows from a previous step. This allows expressing access checks for nested insertions (which otherwise, don't have a way to express a `where` clause). In this implementation we compute the filtering condition on parent based on the nested elements predicate and use it in the `TransactionStep::Filter` step.

Also rearrange ownership tests using folders and add a few additional tests.