Limit nested one-to-many update to only one level
If we have a model such as Venue->[Concert]->[Ticket], we allowed updating the Concert and Ticket models in the same mutation.
This creates a complex API (and we had a bug where we incorrectly processed such a model leading to a model building failure).

In this commit we limit the nested update to only one nested level.

Fixes #871