Support user-assignable primary keys (#881)
Currently, we assume that the primary key is always system-assigned. However, this limits the use
cases, where the user wants to assign the primary key themselves (for example, one obtained through
an external auth system's). This commit adds support for user-assignable primary keys by assuming
that unless the primary key is auto-incrementing or generated UUID, it is user assignable.