Use a distinct favicon for the graphiql app (#879)
So far, we used the same favicon for the graphiql app as for the exograph.dev website. which makes it
hard to distinguish between open tabs for docs vs playground.

In this commit, we add a small "Play" button (matching the one in GraphiQL) over the our logo. We also switch to using a PNG file instead of ICO. PNG has a [pretty wide support](https://caniuse.com/link-icon-png); SVG would have been nicer, but [Safari doesn't support it](https://caniuse.com/link-icon-svg).