Fix coercion of null values in context expressions
If the context value is `null` (for example, due to a missing JWT claim or missing header), we can
return it as is for any type. This allows to correctly handle expressions such as `<something> ||
SomeContext.role == "admin"`, and the `SomeContext.role` is null. In this case, the
`SomeContext.role == "admin"` will evaluate to `false`, and the `||` operator will return the value
of `<something>`.

This will also make it work correctly in the presence of a short-circuiting logic for || and &&.