Fix coercion of null values in context expressions (#887)
If the context value is `null` (for example, due to a explicit null value in JWT claim), we can return it as is for any type. This allows to correct handling of expressions such as `<something> || SomeContext.role == "admin"` when `SomeContext.role` is null. In this case, the `SomeContext.role == "admin"` will evaluate to `false`, and the `||` operator will return the value of `<something>`.

This will also make it work correctly in the presence of a short-circuiting logic for || and &&.