Drop cross-building from CI
Building aarch64-mac-darwin using cross is unreliable and causes us to go through hoops (especially
when it comes to Deno). This commit removes cross-building from CI. Instead, it requires building
aarch64-mac-darwin using a local Mac machine and then uploading the binaries to Github release.

This is a stop-gap measure until Github supports that platform (which is likely to happen soon
https://github.com/github/roadmap/issues/528).

This also fixes #873 by marking the release created by CI as pre-release.