Fix server binding in production
When deploying to fly.io, due to the way we were binding the server (to both "localhost" and "0.0.0.0"), the server failed to start (with a temporary workaround specifying EXO_SERVER_HOST="0.0.0.0").

We now bind the server to either of the addresses based on the deployment mode ("yolo", "dev", or default "prod"). We introduce `EXO_DEPLOYMENT_MODE` env that is set by the `exo` cli to either "yolo" or "dev" when running the respective commands (the default is "prod").