Clear up the disk space in CI (#923)
We have been running out of disk space in CI. 

In this PR, we remove some unneeded directories (Android installation, for example) that come with the default runner. We only do so on non-Windows platforms since we haven't observed the disk space issue on Windows, but if we do, we can implement a similar technique.