Support user-specified value for auto-generated Uuid fields
To support client-generated `Uuid` optionally, this PR now includes a field corresponding to the primary key for creating mutations.

We need to have a formalism for this support. For example, we don't implement a similar mechanism for `autoIncrement()`ed `Int`s, since that is likely to cause confusion when a system-generated value collides with a previous user-specified value.
The difference is, unlike `generate_uuid()`, `autoIncrement()` relies on database state (the underlying sequence). We can deal with it when we address https://github.com/exograph/exograph/issues/926.