Make schema migration idempotent
We had a few issues with the schema migration that are fixed in this commit:
- Handling of non-public schemas was inconsistent
- Handling of blob columns was inconsistent

Also add a schema verification step in integration tests to ensure that the schema migration remains
idempotent.

Also, add the `--apply-to-database` flag to the `exo schema migrate` command to directly apply the
migration to the database.