Fix loading dependencies of an ESM NPM module (#948)
Unlike CommonJS, if an NPM module has ESM sources they are loaded through the same `EmbeddedModuleLoader` pipeline as Deno modules. This would result in us being unable to load dependencies of such a module since they aren't loaded directly from the file system but instead from `EmbeddedModuleLoader` again.

Fixing this requires three steps. First, we have `EmbeddedModuleLoader` use an underlying `NodeResolver` when it is processing a resolution on behalf of a file inside an NPM module (this also allows NPM ESM sources to load Node builtins). Second, we need the ESM sources of transitive dependencies to be in the `EmbeddedModuleLoader`'s map. We achieve this by having the Deno system builder walk the dependencies of NPM modules (stopping when we reach a CommonJS source). Third, we need to pre-generate the ESM wrappers around any CommonJS sources that may be loaded (we originally only handled top-level imports). This is handled by the same logic as step 2, by generating these wrappers for any CommonJS module we encounter while walking the NPM transitive dependencies.

Note that we do not need to walk the dependencies of a CommonJS module itself, since entering a CommonJS module switches the execution mode to have resolution go directly to the virtual filesystem instead of through `EmbeddedModuleLoader`.

With this change, we are able to fully load the `stripe` NPM module and make networked API calls.