Support 'exo playground' command
This command support the use case, where a remote exo server doesn't have introspection enabled
(which is the default behavior in production model), but the user wants to have a good developer
experience with the right schema to execute GraphQL operations against. Here, the schema is fetched
from the local server and GraphQL operations are executed against the remote server.