Support 'exo playground' command (#949)
This command supports the use case where a remote exo server doesn't enable introspection (which is the default behavior in the production model). Still, developers want an excellent experience with the right schema to execute GraphQL operations against. This mode provides that!

In the playground mode, the local server serves schema fetching requests (which only serves introspection queries) and relays non-schema queries to the upstream server. The relaying technique avoids a direct connection from the browser to the upstream server and obviates the need to set up EXO_CORS_DOMAINS in the production server.

Also, show the upstream endpoint URL in the UI when relaying is in effect.