Initialise deno runtime in testing library
Integration tests still suffer from the deno segfault issue since
`exo test` runs multiple tests in parallel and it's not guaranteed
the runtime will be initialized in the main thread.

This change calls deno_core::JsRuntime::init_platform in the testing
crate before it begins to run the tests, making sure that it is
called on the main thread.