Use ctor to ensure deno runtime is initialized early in tests
When running `cargo test` we still get segfaults due to deno_core
issue #247. By using ctor we can initialize the runtime in the lib.rs
file when the test configuration is enabled.

This has been added to both the exo-deno and testing crates, which
is enough to prevent the crash when running standard rust tests.