Do not consider indexed due to unique constraint
Setting up a unique constraint leads to a unique index on those columns, so we don't need to do it
ourselves. Fixes a migration issue where the new schema (correctly) doesn't add an index, but
reconstructing schema from a real database considers that to be an index. So this leads to an
attempt to drop such an index, which fails the migration.