Add to test to ensure multi-mutation transactions
The test ensures that when we execute a multi-mutation operation, Exograph executes them in a transaction, and if any of the mutations fail, it rolls back the entire transaction.

We have another test (transaction-rollbacks/rollback) along the same lines, but that involves executing multiple mutations from a Deno function. This test targets directly called mutations.

Fixes #388