Namespace unique constraint per type (#1001)
Earlier, we assumed that the unique names (to the `unique()` annotations) would be globally unique, which was overly restrictive (and would have interfered with a solution for #945). This change namespaces each name with the type name.

With this change, the following is now valid:

```exo
@postgres
module RsvpModule {
  type Rsvp {
    ...
    @unique("eventRsvp") event: String
    @unique("eventRsvp")  user: User
  }

  type InternalRsvp {
    ...
    @unique("eventRsvp") event: String
    @unique("eventRsvp")  user: User
  }

  ...
}
```

Also, change the model element names to match the naming convention.