Export Deno ops in exo_execution via esm in the extension
As with the tests in DenoModule, later changes in deno mean that use
code can't call op functions directly. Instead these are exported using
ESM from the extension used to set up the ops.

The existing shim code then calls this ESM code. This is unwieldy, with
multiple JS files essentially doing the same thing, so we should look
to improve how it is done in future, ideally removing the shim files
and working directly with the extension.