Update deno patch version in Cargo.lock
The deno patched_1_41_3 branch was changed to remove the
include_js_files_for_snapshotting feature from deno_core. This was
polluting our build, resulting in paths to the Deno extension JS files
being included in the exograph cli executable instead of the code itself
being embedded.

This commit also includes a test to check that the extension code is
inlcluded in the binary rather than loaded at runtime.