Allow external systems to access error reports (#1066)
Earlier, we used stderr to report errors, but that is insufficient for external systems that want to access raw data (diagnostics with line/column numbers and a formatted error string) to display in their way. This commit introduces an error-reporting abstraction and a couple of implementations to make this possible.