Perform queries with one roundtrip to the database (#1091)
Use Exograph's fork of `tokio-postgres` from https://github.com/exograph/rust-postgres, which offers a new `query_with_param_types` method that performs queries in one roundtrip to the database. This requires that we track query parameter types, which is the bulk of the changes in this PR.

Also, drop the direct dependency of the `pgvector` crate. Instead, uses casts to let the database handle it