Treat `Set` correctly when enforcing input/output type disjoinment (#1184)
We considered the raw `Set` type to be a top-level type. This prevented declaring a `Set` of different underlying
types as input and output types in Deno modules. This commit considers the underlying types of the `Set` type
when enforcing the disjoining of input and output types.

Also fixes the computation of unused types to consider operation arguments (in addition to return types).

Fixes #1183