Fix incorrect table alias with multi-level joins
We computed the table alias for a column path by joining the table alias of the column's parent with the column's name. This was unnecessary (and incorrect since the names used in `join` clauses were not the same as the table aliases in predicates/order by) and caused incorrect table aliasing in some cases.

Fixes #1196