Speed up yolo mode by skipping certificate loading
Loading native certificates takes about 100ms on my machine. Since in
yolo mode uses Postgres through either a Unix socket or Docker, we don't
need to load any certificates.

In this PR, we check if all hosts are in Unix sockets and if so, skip
loading the native certs. Similarly, with Docker, we can set the
`sslmode` to `disable` (we already do not load native certs if SSL is
disabled).