Produce diagnostic for many-to-many relationships (#1232)
Currently, we don't support direct (i.e., without a linking type)
many-to-many relationships. This change produces a diagnostic
when such a relationship is detected.

For example, for the following model:
```exo
@postgres
module ConcertDatabase {
  type Concert {
    @pk id: Int = autoIncrement()
    name: String
    artists: Set<Artist>
  }

  type Artist {
    @pk id: Int = autoIncrement()
    name: String
    concerts: Set<Concert>
  }
}
```

`exo build` now produces the following error:

```
error[C000]: Many-to-many relationships without a linking type are not supported. Consider adding a type such as 'ArtistConcert' to connect 'Concert' and 'Artist'
  --> src/index.exo:40:5
   |
6  |     artists: Set<Artist>
   |     ^^^^^^^^^^^^^^^^^^^^
error[C000]: Many-to-many relationships without a linking type are not supported. Consider adding a type such as 'ArtistConcert' to connect 'Artist' and 'Concert'
  --> src/index.exo:46:5
   |
12 |     concerts: Set<Concert>
   |     ^^^^^^^^^^^^^^^^^^^^^^
```