Support `contains` function in access expressions
This allows expressing the dual of the `in` operator in access expressions.
For example, you can express the following rule:

```exo
@access(AuthContext.roles.contains("admin"))
```

This has the same effect as:
```exo
@access("admin" in AuthContext.roles)
```

But a few developers find the `contains` function more intuitive.

Currently, contains supports expression only on context fields.
We can expand this to support other expressions in the future.

Fixes #1244