Make `Router` object-safe
By making `Router` object-safe, we can use it as a trait object in
`CompositeRouter` in a much simpler way (and remove dependency on
specific concrete router types).

Also, rename `ApiRouter` to `Router` to reflect that it is useful
for more than just API routing (specifically, we use it to serve
the playground assets).