Upgrade to http-1.x
Actix uses http-0.2. So far, we used that version
in the rest of the system. However, during Deno upgrade,
we faced issues related to `rustls``, which in turn, stem
from the use of an older `request`, but upgrading that brings
in http-1.

This PR upgrades http to "1" and apply mapping from http-0.2
used by Actix. Once Actix 5.x is released (which uses http-1.x),
we can remove this mapping.