Upgrade to http-1.x and bridge Acitx impl to deal with two versions (#1310)
Actix uses http-0.2. So far, we used that version in the rest of the system. However, during the Deno upgrade, we faced issues related to `rustls``, which in turn, stems from using an older `request`, but upgrading that brings in http-1.

This PR upgrades http to "1" and applies mapping from http-0.2 used by Actix. Once Actix 5.x is released (which will use http-1), we can remove this mapping.