Upgrade Deno to 1.45.5 (#1316)
Needed to upgrade reqwest along the way and use the `ring` feature of `rustls`. Since Deno uses `ring` and if we bring `aws_lc_rs` (the default), initialization of `rustls` fails.

Also needed to use a newer version (our fork) of `oidc_jwt_validator` to use the matching version of `reqwest`.