Implement the infrastructure to support REST (#1341)
Set up the structure to support REST (but not yet implement the leaf resolve functionality for it).

In this PR, builder and resolver/router have been refactored to consider the possibility of multiple protocols (GraphQL and REST). The REST router is enabled only if the `EXO_UNSTABLE_ENABLE_REST_API` is set to `true`.