Support untracked tables (#1428)
In some cases, developers want to use a table or view that already exists in your database. For example, they may want to use a view or a foreign table you created in the database. This PR supports this use case using the `tracked=false` attribute of the `@table` annotation.

For example, to use the `product_profits` view, they can define the following Exograph module:

```exo
@postgres
module TodoDatabase {
  ...

  @table(tracked=false)
  @access(query=true, mutation=false)
  type ProductProfit {
    @pk id: Int
    name: String
    salePrice: Float
    purchasePrice: Float
    profit: Float
  }
}
```

Here, we assume that developers have a matching `product_profits` view in the database:

With this PR, Exograph will map the `ProductProfit` type to the `product_profits` view (it could also be a table). However, Exograph will ignore the `ProductProfit` type for schema migration. 

Exograph will apply the usual access control and infer queries for the `ProductProfit` type as usual including aggregated queries.

Currently, untracked types must set `mutation=false` and must have a field as the primary key. We will lift this restriction in the future.