Simplify migration and make it extensible
Our migration tests used strings in Rust files (for exo and SQL), which required string escaping, prefix stripping, and other modifications. This made reading the test code difficult.
Also, we relied on exact string matches, which required paying too much attention to the precise format.

This refactoring:
1. Move .exo and .sql files outside of rust code
2. Uses `sqlparser` to match without considering the format

Note that in some cases, `sqlparser` doesn't parse valid syntax, so we rely on exact string comparison as a fallback/optimization. We will contribute fixes to `sqlparser` and then remove this reliance.