Introduce the @relation annotation
This removes the overloaded (and confusing) semantics of the `@column` annotation. Instead of using `@column` on the collection side (the primary source of confusion since such annotation is meant to denote the column in the referred type), a dedicated `@relation` annotation is available. This change, along with the earlier change (to make @column not needed in many cases), requires the use of `@relation` only for multiple fields of the same type.

```exo
@postgres
module ConcertDatabase {
  @access(true)
  type Concert {
    @pk id: Int = autoIncrement()
    title: String
    mainVenue: Venue
    altVenue: Venue?
  }

  @access(true)
  type Venue {
    @pk id: Int = autoIncrement()
    name: String
    @relation("mainVenue") mainConcerts: Set<Concert>?
    @relation("altVenue") altConcerts: Set<Concert>?
  }
}
```