Support access control that require a precheck for create/update (#1469)
Some access control expressions require checking a condition with the preexisting data (and not just used-provided arguments).

Consider the following example. When creating an issue, the `self.assignee.position == "developer" part requires querying the database to get the `position` of the employee being assigned to the issue. Similarly, updating an issue requires checking the position of the employee.

```exo
context AuthContext {
  @jwt("sub") id: Int
  @jwt role: String
}

@postgres
module IssueDatabase {
  @access(
    mutation=
      (AuthContext.role == "manager" && self.assignee.position == "developer") ||
      (self.assignee.id == AuthContext.id && self.assignee.position == "developer"),
    ...
  )
  type Issue {
    @pk id: Int = autoIncrement()
    title: String
    assignee: Employee
  }

  @access(true)
  type Employee {
    @pk id: Int = autoIncrement()
    name: String
    position: String
    issues: Set<Issue>?
  }
}
```

In this PR, we introduce `PrecheckAccessPrimitiveExpression` and the associated builder and solver to first perform a check on input (like the `InputAccessPrimitiveExpression` used to, which is now superseded by the introduced `PrecheckAccessPrimitiveExpression`). If the input alone cannot fully evaluate access control, it produces a database query to precheck before executing the primary operation.