Support invariants in integration tests (#1471)
We often want to ensure that the system remains in the same state as before the operation, especially when testing access control rejection with mutations. While this can be done using multi-stage tests, it gets cumbersome.

We now support invariants and operations executed before and after the test.

```
invariants:
  - path: system-state.gql
operation: |
  mutation {
      createProduct(data: {name: "P1", price: 100}) {
          id
      }
  }
auth: |
  {
      "role": "user"
  }
...
```

The format for the invariant file is the same as the init files.